+++ ios/TrueSheetView.mm
@@ -393,6 +393,11 @@ using namespace facebook::react;
     _controller.headerHeight = @(headerHeight);
   }
 
+  CGFloat footerHeight = [_containerView footerHeight];
+  if (footerHeight > 0) {
+    _controller.footerHeight = @(footerHeight);
+  }
+
   if (_eventEmitter) {
     [TrueSheetLifecycleEvents emitMount:_eventEmitter];
   } else {
@@ -580,6 +585,8 @@ using namespace facebook::react;
 }
 
 - (void)containerViewFooterDidChangeSize:(CGSize)newSize {
+  _controller.footerHeight = @(newSize.height);
+  [self setupSheetDetentsForSizeChange];
   [_controller setupAccessibilityContainer];
 }